backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
flask-socketio==5.3.6
flask-limiter==3.5.0
flask-talisman==1.0.0
python-dotenv==1.0.0
tensorflow==2.12.0
pymongo==4.5.0
bcrypt==4.0.1
python-dateutil==2.8.2